Move cookie banner script to a client side rendered component
