+++ .github/actions/deploy/action.yaml
@@ -68,7 +68,7 @@ runs:
         ENVIRONMENT: ${{ inputs.environment }}
         NEXT_PUBLIC_ENVIRONMENT: ${{ inputs.environment }}
         PATH_TO_GITBOOK_DOCS: ${{ inputs.path_to_gitbook_docs }}
-        COOKIE_DOMAIN_SCRIPT: ${{ inputs.cookie_domain_script }}
+        NEXT_PUBLIC_COOKIE_DOMAIN_SCRIPT: ${{ inputs.cookie_domain_script }}
         ALLOW_CRAWLER: ${{ inputs.allow_crawler }}
         NEXT_TELEMETRY_DISABLED: "1"
         NEXT_PUBLIC_COGNITO_REGION: ${{ inputs.aws_region }}
