+++ apps/nextjs-website/.env.default
@@ -1,7 +1,7 @@
 PATH_TO_GITBOOK_DOCS="/<full-path-to-repository-directory>/apps/nextjs-website/docs"
-COOKIE_DOMAIN_SCRIPT="a8c87faf-1769-4c95-a2e5-a6fff26eadab-test"
 ALLOW_CRAWLER="false"
 NEXT_PUBLIC_COGNITO_USER_POOL_WEB_CLIENT_ID=""
 NEXT_PUBLIC_COGNITO_REGION=""
 NEXT_PUBLIC_COGNITO_USER_POOL_ID=""
 NEXT_PUBLIC_ENVIRONMENT="dev"
+NEXT_PUBLIC_COOKIE_DOMAIN_SCRIPT="a8c87faf-1769-4c95-a2e5-a6fff26eadab-test"
