+++ apps/nextjs-website/src/config.ts
@@ -1,6 +1,6 @@
 // TODO: Add environment parser
 export const docsPath = process.env.PATH_TO_GITBOOK_DOCS;
-export const cookieDomainScript = process.env.COOKIE_DOMAIN_SCRIPT;
+export const cookieDomainScript = process.env.NEXT_PUBLIC_COOKIE_DOMAIN_SCRIPT;
 export const environment = process.env.ENVIRONMENT;
 export const docsAssetsPath = '/gitbook/docs';
 export const allowCrawler = process.env.ALLOW_CRAWLER === 'true';
